port atom type to 64bit, 32bit is just so 1990s

+++ CoreLibrary/base.h
@@ -77,9 +77,9 @@ class core_dll _Object {
     friend class _P;
 protected:
 #ifdef ARCH_32
-    uint32 __vfptr_padding_Object;
+    uint64 __vfptr_padding_Object;
 #endif
-    int32 volatile refCount;
+    int64 volatile refCount;
     _Object();
 public:
     virtual ~_Object();
